FETURE/MINOR: Allow to make table "log" ineffective but maintain API
Sometimes you don't want people having access to the livestatus socket
to cause IO, load or memory bumps on a server because they explore the
livestatus API. Querying the "log" table may stress the server by loading
huge log files and there is no way to prevent that with the current
configuration options.

Basically this patch allows to kind of disable the "log" table when
configuration parameter max_cached_messages is set to 0. The table will
still be available but won't return any row, because no log file will be
scanned.

The following mentions "Naemon" because the patch was written for
naemon-livestatus, but it is applicable to Nagios and the original
livestatus broker from Mathias Kettner.

The problem with this new "log" table is that it opens every log file,
naemon.log and all files found in the archives directory. First, when
Naemon is configured to use syslog only, there is no naemon.log file and
an illegitimate error message is logged. Also, when logrotate is used
with the compress option instead of the native Naemon log rotate method,
livestatus tries to open compressed files and emit unwanted log
messages.

Because TableLog is implemented as a member of the Store class, and
because updateLogfileIndex() is called from the TableLog constructor,
this patch seams to be a good way to disable log files scanning. Another
good point is that it won't break the API since the "log" table is still
available; it will just never return any data.